Minor fixes/cleanup

+++ src/primecoinMiner/prime.cpp
@@ -55,7 +55,7 @@ void GeneratePrimeTable(unsigned int nSieveSize)
    for (unsigned int n = 2; n < nPrimeTableLimit; n++)
       if (!vfComposite[n])
          vPrimes.push_back(n);
-   printf("GeneratePrimeTable() : prime table [1, %d] generated with %lu primes\n", nPrimeTableLimit, vPrimes.size());
+   printf("GeneratePrimeTable() : prime table [1, %d] generated with %d primes\n", nPrimeTableLimit, (unsigned int)vPrimes.size());
    vPrimesSize = vPrimes.size();  
 }
